+++ routes/products.js
@@ -2,6 +2,7 @@ const express = require('express');
 const router = express.Router(); //this will called the get, post etc method
 const { Product, validate } = require('../models/product');
 const { Category } = require('../models/category');
+const { isValidObjectId } = require('mongoose');
 
 //for getting all product
 router.get('/', async (req, res) => {
@@ -11,6 +12,8 @@ router.get('/', async (req, res) => {
 
 //for getting one product by id
 router.get('/:id', async (req, res) => {
+    if (!isValidObjectId(req.params.id)) return res.status(400).send("Object id is wrong");
+
     const product = await Product.findById(req.params.id);
     if (!product) return res.status(404).send("The product with the given Id is not found");
     res.send(product);
@@ -18,6 +21,7 @@ router.get('/:id', async (req, res) => {
 
 //delete a specific element with id
 router.delete('/:id', async (req, res) => {
+    if (!isValidObjectId(req.params.id)) return res.status(400).send("Object id is wrong");
     //if id exits or not
     const product = await Product.findByIdAndRemove(req.params.id)
     if (!product) return res.status(404).send("The product with the given Id is not found");
@@ -53,6 +57,7 @@ router.post('/', async (req, res) => {
 
 //for updating the existing data
 router.put('/:id', async (req, res) => {
+    if (!isValidObjectId(req.params.id)) return res.status(400).send("Object id is wrong");
     const { error } = validate(req.body);
     if (error) return res.status(400).send(error.details[0].message);
